Conflicts:
	platform/product/monitor/src/main/resources/application.properties
	platform/product/monitor/src/main/webapp/WEB-INF/layouts/selectMonitorType.jsp

+++ platform/product/monitor/src/main/java/com/sinosoft/one/monitor/application/domain/UrlService.java
@@ -1,16 +1,22 @@
 package com.sinosoft.one.monitor.application.domain;
 
 import com.sinosoft.one.monitor.application.model.BizScenario;
+import com.sinosoft.one.monitor.application.model.EumUrl;
 import com.sinosoft.one.monitor.application.model.Url;
 import com.sinosoft.one.monitor.application.repository.EumUrlRepository;
 import com.sinosoft.one.monitor.application.repository.UrlRepository;
+import com.sinosoft.one.monitor.common.ResourceType;
+import com.sinosoft.one.monitor.resources.domain.ResourcesService;
+import com.sinosoft.one.monitor.resources.model.Resource;
 import com.sinosoft.one.monitor.resources.repository.ResourcesRepository;
+import com.sinosoft.one.monitor.utils.CurrentUserUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -32,12 +38,48 @@ public class UrlService {
 
 	@Autowired
 	private EumUrlRepository eumUrlRepository;
+
+    @Autowired
+    private ResourcesService resourcesService;
     /**
      * 新增一个URL.
      */
     @Transactional(readOnly = false)
     public void saveUrl(Url url) {
         urlRepository.save(url);
+    }
+
+    @Transactional(readOnly = false)
+    public void saveUrlWithTransactional(Url url,BizScenario bizScenario){
+        EumUrl eumUrl=new EumUrl();
+        eumUrl.setUrl(url.getUrl());
+        eumUrl.setApplication(bizScenario.getApplication());
+        //获得当前用户id
+        /*String creatorId = CurrentUserUtil.getCurrentUser().getId();*/
+        //如果新增加的url是库中没有的，那么入库
+        //当前Url所属的业务场景
+        url.setBizScenario(bizScenario);
+        url.setStatus(String.valueOf(1));
+        //保存当前创建url的用户
+        url.setCreatorId(CurrentUserUtil.getCurrentUser().getId());
+        //开发阶段固定用户id
+        /*url.setCreatorId("4028921a3cfba342013cfba4623e0000");*/
+        url.setCreateTime(new Date());
+        saveUrl(url);
+        //向EUM_URL表中插入记录（url的application信息）
+        eumUrl.setUrlId(url.getId());
+        eumUrl.setRecordTime(new Date());
+        eumUrlRepository.save(eumUrl);
+        saveResourceWithUrl(url);
+    }
+
+    private void saveResourceWithUrl(Url url){
+        //资源表存入新建业务场景的信息
+        Resource resource=new Resource();
+        resource.setResourceId(url.getId());
+        resource.setResourceName(url.getDescription());
+        resource.setResourceType(ResourceType.APPLICATION_SCENARIO_URL.name());
+        resourcesService.saveResource(resource);
     }
 
     /**
@@ -55,15 +97,21 @@ public class UrlService {
     public void deleteUrl(String bizScenarioId, String id) {
 	    //先删除中间表GE_MONITOR_BIZ_SCENARIO_URL的记录
 	    urlRepository.deleteBizScenarioAndUrl(bizScenarioId,id);
-	    //先删除中间表GE_MONITOR_URL_METHOD的记录
-	    urlRepository.deleteUrlAndMethod(id);
-	    //删除GE_MONITOR_URL的记录
-	    urlRepository.delete(id);
-	    //删除Resources表中的记录
-	    resourcesRepository.delete(resourcesRepository.findByResourceId(id));
-	    // 删除eumUrl表中的记录
-	    eumUrlRepository.delete(eumUrlRepository.findByUrlId(id));
-        urlRepository.delete(id);
+        //如果当前url有关联的其它业务场景，那么不能删除
+        List<String> bizScenarioIds=urlRepository.selectUrlsWithUrlId(id);
+        if(null!=bizScenarioIds&&bizScenarioIds.size()>0){
+            return;
+            //如果当前url没有关联其它的业务场景，那么删除当前url
+        }else if(null==bizScenarioIds||bizScenarioIds.size()==0){
+            //先删除中间表GE_MONITOR_URL_METHOD的记录
+            urlRepository.deleteUrlAndMethod(id);
+            //删除Resources表中的记录
+            resourcesRepository.delete(resourcesRepository.findByResourceId(id));
+            // 删除eumUrl表中的记录
+            eumUrlRepository.delete(eumUrlRepository.findByUrlId(id));
+            //删除GE_MONITOR_URL的记录
+            urlRepository.delete(id);
+        }
     }
 
     /**
